+++ src/array/bytes_representation.rs
@@ -1,12 +1,26 @@
-use derive_more::{Display, From};
+use derive_more::Display;
 
 /// The decoded representation of `bytes`.
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Display, From)]
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Display)]
 pub enum BytesRepresentation {
-    /// The output size is known.
-    #[display(fmt = "bytes: {_0}")]
-    KnownSize(u64),
-    /// The output size may vary.
-    #[display(fmt = "bytes: variable")]
-    VariableSize,
+    /// The output size is fixed.
+    #[display(fmt = "fixed size: {_0}")]
+    FixedSize(u64),
+    /// The output size is bounded.
+    #[display(fmt = "bounded size: {_0}")]
+    BoundedSize(u64),
+    /// The output size is unbounded/indeterminate.
+    #[display(fmt = "unbounded size")]
+    UnboundedSize,
+}
+
+impl BytesRepresentation {
+    /// Return the fixed or bounded size of the bytes representations, or [`None`] if the size is unbounded.
+    #[must_use]
+    pub fn size(&self) -> Option<u64> {
+        match self {
+            Self::FixedSize(size) | Self::BoundedSize(size) => Some(*size),
+            Self::UnboundedSize => None,
+        }
+    }
 }