+++ src/array/codec/array_to_bytes/zfp/zfp_partial_decoder.rs
@@ -30,7 +30,7 @@ impl ArrayPartialDecoderTraits for ZfpPartialDecoder<'_> {
     ) -> Result<Vec<Vec<u8>>, CodecError> {
         let encoded_value = self
             .input_handle
-            .decode(&BytesRepresentation::VariableSize)?;
+            .decode(&BytesRepresentation::UnboundedSize)?; // FIXME: Fixed/bounded?
         let mut out = Vec::with_capacity(decoded_regions.len());
         match encoded_value {
             Some(encoded_value) => {