+++ src/array/codec/bytes_to_bytes/gzip.rs
@@ -51,7 +51,7 @@ mod tests {
     fn codec_gzip_round_trip1() {
         let elements: Vec<u16> = (0..32).collect();
         let bytes = safe_transmute::transmute_to_bytes(&elements).to_vec();
-        let bytes_representation = BytesRepresentation::KnownSize(bytes.len() as u64);
+        let bytes_representation = BytesRepresentation::FixedSize(bytes.len() as u64);
 
         let configuration: GzipCodecConfiguration = serde_json::from_str(JSON_VALID).unwrap();
         let codec = GzipCodec::new_with_configuration(&configuration);
@@ -65,7 +65,7 @@ mod tests {
     fn codec_gzip_partial_decode() {
         let elements: Vec<u16> = (0..8).collect();
         let bytes = safe_transmute::transmute_to_bytes(&elements).to_vec();
-        let bytes_representation = BytesRepresentation::KnownSize(bytes.len() as u64);
+        let bytes_representation = BytesRepresentation::FixedSize(bytes.len() as u64);
 
         let configuration: GzipCodecConfiguration = serde_json::from_str(JSON_VALID).unwrap();
         let codec = GzipCodec::new_with_configuration(&configuration);