+++ src/array/codec/bytes_to_bytes/gzip/gzip_codec.rs
@@ -111,8 +111,18 @@ impl BytesToBytesCodecTraits for GzipCodec {
 
     fn compute_encoded_size(
         &self,
-        _decoded_representation: &BytesRepresentation,
+        decoded_representation: &BytesRepresentation,
     ) -> BytesRepresentation {
-        BytesRepresentation::VariableSize
+        match decoded_representation.size() {
+            Some(size) => {
+                // https://www.gnu.org/software/gzip/manual/gzip.pdf
+                const HEADER_TRAILER_OVERHEAD: u64 = 10 + 8; // TODO: validate that extra headers are not populated
+                const BLOCK_SIZE: u64 = 32768;
+                const BLOCK_OVERHEAD: u64 = 5;
+                let blocks_overhead = BLOCK_OVERHEAD * ((size + BLOCK_SIZE - 1) / BLOCK_SIZE);
+                BytesRepresentation::BoundedSize(size + HEADER_TRAILER_OVERHEAD + blocks_overhead)
+            }
+            None => BytesRepresentation::UnboundedSize,
+        }
     }
 }